feat: refactor game lobby and search results pages
